# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "Favorite": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "place": {
            "$ref": "#/components/schemas/Place"
          }
        },
        "required": [
          "id",
          "label",
          "place",
          "created_at"
        ],
        "title": "Favorite",
        "type": "object"
      },
      "FavoriteCreate": {
        "additionalProperties": false,
        "properties": {
          "label": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Label",
            "type": "string"
          },
          "place": {
            "$ref": "#/components/schemas/Place"
          }
        },
        "required": [
          "label",
          "place"
        ],
        "title": "FavoriteCreate",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HistoryEntry": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "destination": {
            "$ref": "#/components/schemas/Place"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "mode": {
            "$ref": "#/components/schemas/TravelMode"
          },
          "origin_label": {
            "title": "Origin Label",
            "type": "string"
          }
        },
        "required": [
          "id",
          "origin_label",
          "destination",
          "mode",
          "created_at",
          "expires_at"
        ],
        "title": "HistoryEntry",
        "type": "object"
      },
      "Locale": {
        "enum": [
          "he",
          "en"
        ],
        "title": "Locale",
        "type": "string"
      },
...
      "MobilityResponse": {
        "additionalProperties": false,
        "properties": {
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "vehicles": {
            "items": {
              "$ref": "#/components/schemas/MobilityVehicle"
            },
            "title": "Vehicles",
            "type": "array"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "vehicles",
          "generated_at"
        ],
        "title": "MobilityResponse",
        "type": "object"
      },
...
      "RoutePlanRequest": {
        "additionalProperties": false,
        "properties": {
          "accessibility": {
            "$ref": "#/components/schemas/AccessibilityPreferences"
          },
          "arrive_by": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Arrive By"
          },
          "constraints": {
            "$ref": "#/components/schemas/RouteConstraints"
          },
          "depart_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Depart At"
          },
          "destination": {
            "$ref": "#/components/schemas/Coordinate"
          },
          "include_comparisons": {
            "default": true,
            "title": "Include Comparisons",
            "type": "boolean"
          },
          "locale": {
            "$ref": "#/components/schemas/Locale",
            "default": "he"
          },
          "mode": {
            "$ref": "#/components/schemas/TravelMode",
            "default": "walk"
          },
          "origin": {
            "$ref": "#/components/schemas/Coordinate"
          },
          "preference": {
            "$ref": "#/components/schemas/RoutePreference",
            "default": "fastest"
          },
          "vehicle": {
            "$ref": "#/components/schemas/VehicleProfile"
          }
        },
        "required": [
          "origin",
          "destination"
        ],
        "title": "RoutePlanRequest",
        "type": "object"
      },
      "RoutePlanResponse": {
        "additionalProperties": false,
        "properties": {
          "data_version": {
            "title": "Data Version",
            "type": "string"
          },
          "engine_profile": {
            "title": "Engine Profile",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "routes": {
            "items": {
              "$ref": "#/components/schemas/RouteAlternative"
            },
            "title": "Routes",
            "type": "array"
          }
        },
        "required": [
          "request_id",
          "routes",
          "data_version",
          "engine_profile"
        ],
        "title": "RoutePlanResponse",
        "type": "object"
      },
...
      "TravelMode": {
        "enum": [
          "walk",
          "bike",
          "scooter",
          "car",
          "motorcycle",
          "truck",
          "transit",
          "bike_transit",
          "scooter_transit",
          "rental_transit"
        ],
        "title": "TravelMode",
        "type": "string"
      },
      "UserPreferences": {
        "additionalProperties": false,
        "properties": {
          "default_mode": {
            "$ref": "#/components/schemas/TravelMode",
            "default": "walk"
          },
          "default_walk_preference": {
            "$ref": "#/components/schemas/RoutePreference",
            "default": "balanced_shade"
          },
          "history_enabled": {
            "default": false,
            "title": "History Enabled",
            "type": "boolean"
          },
          "locale": {
            "$ref": "#/components/schemas/Locale",
            "default": "he"
          }
        },
        "title": "UserPreferences",
        "type": "object"
      },
...
    }
  },
  "info": {
    "summary": "Israel-wide shaded, multimodal, and low-signal navigation",
    "title": "Naviz API",
    "version": "0.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/me/favorites": {
      "get": {
        "operationId": "list_favorites_v1_me_favorites_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Favorite"
                  },
                  "title": "Response List Favorites V1 Me Favorites Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Favorites",
        "tags": [
          "account"
        ]
      },
      "post": {
        "operationId": "save_favorite_v1_me_favorites_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FavoriteCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Favorite"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Favorite",
        "tags": [
          "account"
        ]
      }
    },
    "/v1/me/favorites/{favorite_id}": {
      "delete": {
        "operationId": "delete_favorite_v1_me_favorites__favorite_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "favorite_id",
            "required": true,
            "schema": {
              "title": "Favorite Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Favorite",
        "tags": [
          "account"
        ]
      }
    },
    "/v1/me/history": {
      "get": {
        "operationId": "get_history_v1_me_history_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/HistoryEntry"
                  },
                  "title": "Response Get History V1 Me History Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get History",
        "tags": [
          "account"
        ]
      }
    },
    "/v1/me/preferences": {
      "get": {
        "operationId": "get_preferences_v1_me_preferences_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPreferences"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Preferences",
        "tags": [
          "account"
        ]
      },
      "put": {
        "operationId": "update_preferences_v1_me_preferences_put",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserPreferences"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPreferences"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Preferences",
        "tags": [
          "account"
        ]
      }
    },
    "/v1/mobility/vehicles": {
      "get": {
        "operationId": "mobility_vehicles_v1_mobility_vehicles_get",
        "parameters": [
          {
            "in": "query",
            "name": "min_latitude",
            "required": true,
            "schema": {
              "maximum": 90,
              "minimum": -90,
              "title": "Min Latitude",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "min_longitude",
            "required": true,
            "schema": {
              "maximum": 180,
              "minimum": -180,
              "title": "Min Longitude",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "max_latitude",
            "required": true,
            "schema": {
              "maximum": 90,
              "minimum": -90,
              "title": "Max Latitude",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "max_longitude",
            "required": true,
            "schema": {
              "maximum": 180,
              "minimum": -180,
              "title": "Max Longitude",
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobilityResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mobility Vehicles",
        "tags": [
          "mobility"
        ]
      }
    },
    "/v1/routes/plan": {
      "post": {
        "operationId": "plan_route_v1_routes_plan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoutePlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutePlanResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Plan Route",
        "tags": [
          "routing"
        ]
      }
    },
    "/v1/routes/reroute": {
      "post": {
        "operationId": "reroute_v1_routes_reroute_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RerouteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutePlanResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reroute",
        "tags": [
          "routing"
        ]
      }
    },
    "/v1/search": {
      "get": {
        "operationId": "search_v1_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "maxLength": 120,
              "minLength": 0,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "language",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Locale",
              "default": "he"
            }
          },
          {
            "in": "query",
            "name": "latitude",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 90,
                  "minimum": -90,
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Latitude"
            }
          },
          {
            "in": "query",
            "name": "longitude",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 180,
                  "minimum": -180,
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Longitude"
            }
          },
          {
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "in": "query",
            "name": "bbox",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Bbox"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 8,
              "maximum": 20,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search",
        "tags": [
          "search"
        ]
      }
    },
    "/v1/search/reverse": {
      "get": {
        "operationId": "reverse_search_v1_search_reverse_get",
        "parameters": [
          {
            "in": "query",
            "name": "latitude",
            "required": true,
            "schema": {
              "maximum": 90,
              "minimum": -90,
              "title": "Latitude",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "longitude",
            "required": true,
            "schema": {
              "maximum": 180,
              "minimum": -180,
              "title": "Longitude",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "language",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Locale",
              "default": "he"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Place"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reverse Search",
        "tags": [
          "search"
        ]
      }
    },
    "/v1/shadows/scene": {
      "post": {
        "operationId": "shadow_scene_v1_shadows_scene_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShadowSceneRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShadowSceneResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Shadow Scene",
        "tags": [
          "shade"
        ]
      }
    }
  }
}